feat: refactor schemesto be in 1 file
