--- src/resources/boards/schemas/addBoardSchema.js
@@ -1,11 +1,0 @@
-const { boardBody, boardRes } = require('./elementSchemas');
-
-const addBoardSchema = {
-  body: boardBody,
-  tags: ['Board'],
-  response: {
-    201: boardRes,
-  },
-};
-
-module.exports = addBoardSchema;
