--- src/resources/boards/schemas/deleteBoardSchema.js
@@ -1,12 +1,0 @@
-const { typeString } = require('./elementSchemas');
-
-const deleteUserSchema = {
-  params: {
-    boardId: typeString,
-  },
-  tags: ['Board'],
-  response: {
-    204: typeString,
-  },
-};
-module.exports = deleteUserSchema;
