--- src/resources/boards/schemas/getAllBoardsSchema.js
@@ -1,13 +1,0 @@
-const { boardRes } = require('./elementSchemas');
-
-const getAllUsersSchema = {
-  tags: ['Board'],
-  response: {
-    200: {
-      type: 'array',
-      items: boardRes,
-    },
-  },
-};
-
-module.exports = getAllUsersSchema;
