+++ src/resources/boards/schemas/index.js
@@ -1,8 +1,51 @@
-const getAllBoardsSchema = require('./getAllBoardsSchema');
-const getBoardSchema = require('./getBoardSchema');
-const addBoardSchema = require('./addBoardSchema');
-const updateBoardSchema = require('./updateBoardSchema');
-const deleteBoardSchema = require('./deleteBoardSchema');
+const { boardBody, boardRes, typeString } = require('./elementSchemas');
+
+const getAllBoardsSchema = {
+  response: {
+    200: {
+      type: 'array',
+      items: boardRes,
+    },
+  },
+};
+
+const getBoardSchema = {
+  params: {
+    boardId: typeString,
+  },
+  response: {
+    200: boardRes,
+  },
+};
+
+const addBoardSchema = {
+  body: boardBody,
+  tags: ['Board'],
+  response: {
+    201: boardRes,
+  },
+};
+
+const updateBoardSchema = {
+  body: boardBody,
+  params: {
+    boardId: typeString,
+  },
+  tags: ['Board'],
+  response: {
+    200: boardRes,
+  },
+};
+
+const deleteBoardSchema = {
+  params: {
+    boardId: typeString,
+  },
+  tags: ['Board'],
+  response: {
+    204: typeString,
+  },
+};
 
 module.exports = {
   getAllBoardsSchema,
