--- src/resources/boards/schemas/updateBoardSchema.js
@@ -1,14 +1,0 @@
-const { typeString, boardBody, boardRes } = require('./elementSchemas');
-
-const updateBoardSchema = {
-  body: boardBody,
-  params: {
-    boardId: typeString,
-  },
-  tags: ['Board'],
-  response: {
-    200: boardRes,
-  },
-};
-
-module.exports = updateBoardSchema;
